refactor(backend): split Space Skill read bindings
