+++ src/backend/src/agentclaw/community/core/repository/protocols/skill_center.py
@@ -39,8 +39,6 @@ from .skill_center_types import (
     SpaceSkillGrantSetRecord,
     DraftEditLeaseRecord,
     SkillVersionRecord,
-    SpaceSkillVersionRecord,
-    ConsumableSpaceSkillRecord,
     SpaceSkillDraftRecord,
     DraftDeleteRecord,
     DraftUpgradeRecord,
@@ -353,30 +351,6 @@ class SkillVersionMaterializationRepositoryProtocol(Protocol):
         description: str,
         published_at: datetime,
     ) -> PublishedMaterializedSkillVersion: ...
-
-
-@runtime_checkable
-class SpaceSkillVersionReadRepository(Protocol):
-    @abstractmethod
-    def list_published(
-        self, *, space_id: int, skill_id: int, env: str, offset: int, limit: int
-    ) -> tuple[int, list[SpaceSkillVersionRecord]]: ...
-
-    @abstractmethod
-    def get_published_ordinal(
-        self, *, space_id: int, skill_id: int, version: int, env: str
-    ) -> SpaceSkillVersionRecord: ...
-
-    @abstractmethod
-    def list_consumable_candidates(
-        self,
-        *,
-        space_id: int,
-        env: str,
-        keyword: str | None,
-        offset: int,
-        limit: int,
-    ) -> tuple[int, list[ConsumableSpaceSkillRecord]]: ...
 
 
 @runtime_checkable
