+++ src/backend/src/agentclaw/community/di/modules/space_skill_repository_bindings.py
@@ -1,0 +1,54 @@
+"""DI bindings for the Space Skill authoring/read persistence seams."""
+
+from injector import Binder, singleton
+
+from agentclaw.community.core.repository.implementations.skill_center.space_skill import (
+    SpaceSkillRepository as UnifiedSpaceSkillRepository,
+)
+from agentclaw.community.core.repository.implementations.skill_center.space_skill_draft import (
+    SpaceSkillDraftRepository as UnifiedSpaceSkillDraftRepository,
+)
+from agentclaw.community.core.repository.implementations.skill_center.space_skill_read import (
+    SpaceSkillReadRepository as UnifiedSpaceSkillReadRepository,
+)
+from agentclaw.community.core.repository.implementations.skill_center.space_skill_version_read import (
+    SpaceSkillVersionReadRepository as UnifiedSpaceSkillVersionReadRepository,
+)
+from agentclaw.community.core.repository.protocols.skill_center import (
+    DraftEditLeaseRepository,
+    SpaceSkillDraftRepository,
+    SpaceSkillReadRepository,
+    SpaceSkillRepository,
+)
+from agentclaw.community.core.repository.protocols.space_skill_version import (
+    SpaceSkillVersionReadRepository,
+)
+
+
+def bind_space_skill_repositories(binder: Binder) -> None:
+    """Bind the additive Space Skill persistence contracts as singletons."""
+    binder.bind(
+        SpaceSkillRepository,
+        to=UnifiedSpaceSkillRepository,
+        scope=singleton,
+    )
+    binder.bind(
+        SpaceSkillDraftRepository,
+        to=UnifiedSpaceSkillDraftRepository,
+        scope=singleton,
+    )
+    binder.bind(
+        SpaceSkillReadRepository,
+        to=UnifiedSpaceSkillReadRepository,
+        scope=singleton,
+    )
+    binder.bind(
+        SpaceSkillVersionReadRepository,
+        to=UnifiedSpaceSkillVersionReadRepository,
+        scope=singleton,
+    )
+    binder.bind(
+        DraftEditLeaseRepository,
+        to=UnifiedSpaceSkillRepository,
+        scope=singleton,
+    )
